finished the dashboard DAO and rethought where I was going with this and I think i am happier with this new way.
The new way is to use graphs and not pick the first col but pick all cols and plot points for all.  Expectation is that the query only returns one row then this is saved as point in time.

The graph plots all points in time basically value against time.

Will make more sophisticated in later versions but I like this way for what I want it to be for a moment so poo