blowing the unit tests away for the moment (shock horror)
I am going to rebuild them all from scratch post r1-final.  This is because they are too scratty and I need to work on a proper strategy for testing each component rather than the big mess that was there as a test framework

This means that it will now build with Maven without having to skip tests!