loads of fixes around report generation - been testing jasper output and the dreaded tabbed xls
had to upgrade jasper a little and a new version of itext.

Found a bug in commons vfs but just upgraded to v2, changed a few imports and it seems to work fine.  I am going to write a test package to test file output to be sure.

Fixed up the file download bit and made it tabbed xls compatable.