loads of stuff here:
refactored parameter code a bit to fix a bug

fixed:

issue 105 - Testing and existing datasource
issue 103 - Deleting a job with parameters
issue 102 - deleting a group with jobs/dashboards
issue 101 - deleteing a datasource in use

Also got it working with embeded jetty thanks to the jetty console plugin for maven - which is amazing.

You can now start up a whole instance of Dash Reports by:

java -jar dashreports-jetty-console.war 

It will set up a derby database and an admin user with password of password and run it in a jetty container.